+++ business-services/egov-hrms/src/main/java/org/egov/hrms/config/PropertiesManager.java
@@ -93,6 +93,9 @@ public class PropertiesManager {
 	
 	@Value("${kafka.topics.notification.sms}")
 	public String coreNotificationTopic;
+
+	@Value("${kafka.topics.hrms.updateData}")
+	public String updateTopic;
 	
 	
 	//Variables