+++ business-services/egov-hrms/src/main/java/org/egov/hrms/service/EmployeeService.java
@@ -136,10 +136,10 @@ public class EmployeeService {
 	 */
 	public EmployeeResponse search(EmployeeSearchCriteria criteria, RequestInfo requestInfo) {
 		boolean  userChecked = false;
-		if(null == criteria.getIsActive() || criteria.getIsActive())
+		/*if(null == criteria.getIsActive() || criteria.getIsActive())
 			criteria.setIsActive(true);
 		else
-			criteria.setIsActive(false);
+			criteria.setIsActive(false);*/
         Map<String, User> mapOfUsers = new HashMap<String, User>();
 		if(!StringUtils.isEmpty(criteria.getPhone()) || !CollectionUtils.isEmpty(criteria.getRoles())) {
             Map<String, Object> userSearchCriteria = new HashMap<>();
@@ -336,8 +336,8 @@ public class EmployeeService {
 			enrichUpdateRequest(employee, requestInfo, existingEmployees);
 			updateUser(employee, requestInfo);
 		});
-		hrmsProducer.push(propertiesManager.getUpdateEmployeeTopic(), employeeRequest);
-
+		hrmsProducer.push(propertiesManager.getUpdateTopic(), employeeRequest);
+		//notificationService.sendReactivationNotification(employeeRequest);
 		return generateResponse(employeeRequest);
 	}
 	
@@ -511,6 +511,29 @@ public class EmployeeService {
 			});
 
 		}
+		if(employee.getReactivationDetails() != null){
+			employee.getReactivationDetails().stream().forEach(reactivationDetails -> {
+				if(reactivationDetails.getId() == null){
+					reactivationDetails.setId(UUID.randomUUID().toString());
+					reactivationDetails.setAuditDetails(auditDetails);
+					employee.getDocuments().forEach(employeeDocument -> {
+						employeeDocument.setReferenceId(reactivationDetails.getId());
+					});
+				}
+				else{
+					if(!existingEmpData.getReactivationDetails().stream()
+							.filter(reactivationDetails1 -> reactivationDetails1.getId().equals(reactivationDetails.getId()))
+							.findFirst().orElse(null)
+							.equals(reactivationDetails)){
+						reactivationDetails.getAuditDetails().setLastModifiedBy(requestInfo.getUserInfo().getUserName());
+						reactivationDetails.getAuditDetails().setLastModifiedDate(new Date().getTime());
+					}
+				}
+			});
+
+		}
+
+
 	}
 
 	private EmployeeResponse generateResponse(EmployeeRequest employeeRequest) {