+++ business-services/egov-hrms/CHANGELOG.md
@@ -1,5 +1,10 @@
+# Changelog
+All notable changes to this module will be documented in this file.
 
-## 1.1.0 - 2020/05/27
+## 1.2.0 - 2021-01-12
+- Added employee reactivation feature
+
+## 1.1.0 - 2020-05-27
 
 - Upgraded to `tracer:2.0.0-SNAPSHOT`
 - Upgraded to `Spring boot 2.2.6`